+++ ompi/mca/coll/base/coll_base_alltoall.c
@@ -48,7 +48,7 @@
  * and count) to send the data to the other.
  */
 int
-mca_coll_base_alltoall_intra_basic_inplace(const void *rbuf, int rcount,
+mca_coll_base_alltoall_intra_basic_inplace(const void *rbuf, size_t rcount,
                                            struct ompi_datatype_t *rdtype,
                                            struct ompi_communicator_t *comm,
                                            mca_coll_base_module_t *module)
@@ -177,9 +177,9 @@ mca_coll_base_alltoall_intra_basic_inplace(const void *rbuf, int rcount,
     return err;
 }
 
-int ompi_coll_base_alltoall_intra_pairwise(const void *sbuf, int scount,
+int ompi_coll_base_alltoall_intra_pairwise(const void *sbuf, size_t scount,
                                             struct ompi_datatype_t *sdtype,
-                                            void* rbuf, int rcount,
+                                            void* rbuf, size_t rcount,
                                             struct ompi_datatype_t *rdtype,
                                             struct ompi_communicator_t *comm,
                                             mca_coll_base_module_t *module)
@@ -236,9 +236,9 @@ int ompi_coll_base_alltoall_intra_pairwise(const void *sbuf, int scount,
 }
 
 
-int ompi_coll_base_alltoall_intra_bruck(const void *sbuf, int scount,
+int ompi_coll_base_alltoall_intra_bruck(const void *sbuf, size_t scount,
                                          struct ompi_datatype_t *sdtype,
-                                         void* rbuf, int rcount,
+                                         void* rbuf, size_t rcount,
                                          struct ompi_datatype_t *rdtype,
                                          struct ompi_communicator_t *comm,
                                          mca_coll_base_module_t *module)
@@ -375,9 +375,9 @@ int ompi_coll_base_alltoall_intra_bruck(const void *sbuf, int scount,
  *                    - wait for any request to complete
  *                    - replace that request by the new one of the same type.
  */
-int ompi_coll_base_alltoall_intra_linear_sync(const void *sbuf, int scount,
+int ompi_coll_base_alltoall_intra_linear_sync(const void *sbuf, size_t scount,
                                                struct ompi_datatype_t *sdtype,
-                                               void* rbuf, int rcount,
+                                               void* rbuf, size_t rcount,
                                                struct ompi_datatype_t *rdtype,
                                                struct ompi_communicator_t *comm,
                                                mca_coll_base_module_t *module,
@@ -534,9 +534,9 @@ int ompi_coll_base_alltoall_intra_linear_sync(const void *sbuf, int scount,
 }
 
 
-int ompi_coll_base_alltoall_intra_two_procs(const void *sbuf, int scount,
+int ompi_coll_base_alltoall_intra_two_procs(const void *sbuf, size_t scount,
                                              struct ompi_datatype_t *sdtype,
-                                             void* rbuf, int rcount,
+                                             void* rbuf, size_t rcount,
                                              struct ompi_datatype_t *rdtype,
                                              struct ompi_communicator_t *comm,
                                              mca_coll_base_module_t *module)
@@ -613,9 +613,9 @@ int ompi_coll_base_alltoall_intra_two_procs(const void *sbuf, int scount,
 
 /* copied function (with appropriate renaming) starts here */
 
-int ompi_coll_base_alltoall_intra_basic_linear(const void *sbuf, int scount,
+int ompi_coll_base_alltoall_intra_basic_linear(const void *sbuf, size_t scount,
                                                struct ompi_datatype_t *sdtype,
-                                               void* rbuf, int rcount,
+                                               void* rbuf, size_t rcount,
                                                struct ompi_datatype_t *rdtype,
                                                struct ompi_communicator_t *comm,
                                                mca_coll_base_module_t *module)
