+++ ompi/mca/coll/base/coll_base_exscan.c
@@ -32,7 +32,7 @@
  * Returns:   MPI_SUCCESS or error code
  */
 int
-ompi_coll_base_exscan_intra_linear(const void *sbuf, void *rbuf, int count,
+ompi_coll_base_exscan_intra_linear(const void *sbuf, void *rbuf, size_t count,
                                   struct ompi_datatype_t *dtype,
                                   struct ompi_op_t *op,
                                   struct ompi_communicator_t *comm,
@@ -140,7 +140,7 @@ ompi_coll_base_exscan_intra_linear(const void *sbuf, void *rbuf, int count,
  * Limitations: intra-communicators only
  */
 int ompi_coll_base_exscan_intra_recursivedoubling(
-    const void *sendbuf, void *recvbuf, int count, struct ompi_datatype_t *datatype,
+    const void *sendbuf, void *recvbuf, size_t count, struct ompi_datatype_t *datatype,
     struct ompi_op_t *op, struct ompi_communicator_t *comm,
     mca_coll_base_module_t *module)
 {
