+++ ompi/mca/coll/basic/coll_basic_alltoallv.c
@@ -42,9 +42,9 @@
  *	Returns:	- MPI_SUCCESS or an MPI error code
  */
 int
-mca_coll_basic_alltoallv_inter(const void *sbuf, const int *scounts, const int *sdisps,
+mca_coll_basic_alltoallv_inter(const void *sbuf, const size_t *scounts, const ptrdiff_t *sdisps,
                                struct ompi_datatype_t *sdtype, void *rbuf,
-                               const int *rcounts, const int *rdisps,
+                               const size_t *rcounts, const ptrdiff_t *rdisps,
                                struct ompi_datatype_t *rdtype,
                                struct ompi_communicator_t *comm,
                                mca_coll_base_module_t *module)
